// src/app/interface/person.interface.ts
export interface Person{
    id: number;
    firstName: string;
    lastName: string;
    displayName: string;
    email: string;
    primaryAddress: string;
    secondaryAddress: string;
    jobTitle: string;
    jobDescription: string;
    country: string;
    city: string;
    state: string;
    zip: number;
    isFavorite: boolean;
    avatar: string;
    phone: string;
    personNotes: string;
    date: string;
}